refactor(utils): relocate collateralPreviewUtils to shared utils directory
